Hot middleware and webpack: still need to make it work though
