+++ src/.babelrc
@@ -1,0 +1,11 @@
+{
+  "presets": ["es2015"],
+  "env": {
+    "development": {
+      "presets": ["react-hmre"]
+    }
+  },
+  "plugins": [
+    "babel-root-import"
+  ]
+}
