+++ src/public/js/scatterPlot.js
@@ -12,185 +12,185 @@
  ******************************************************************************/
 
 function Scatter(args){
-	var containerID = args.ID || "body",
-		containerWidth = args.width || 900,
-		containerHeight = args.height || 900,
-		margin = args.margin || {top: 10, right: 30, bottom: 10, left: 10},
-		yData = args.yData,
-		xData = args.xData,
-		label = args.label,
-		sort =  args.sort,
-		clickCallBack = args.clickCallBack;	
+    var containerID = args.ID || "body",
+	containerWidth = args.width || 900,
+	containerHeight = args.height || 900,
+	margin = args.margin || {top: 10, right: 30, bottom: 10, left: 10},
+	yData = args.yData,
+	xData = args.xData,
+	label = args.label,
+	sort =  args.sort,
+	clickCallBack = args.clickCallBack;	
 
-	// console.log("in scatter", label, sort);
+    // console.log("in scatter", label, sort);
 
-	var width = containerWidth - margin.left - margin.right;
-	var height = containerHeight - margin.top - margin.bottom;
+    var width = containerWidth - margin.left - margin.right;
+    var height = containerHeight - margin.top - margin.bottom;
 
-	//find min and max
-	var yMin = Number.MAX_SAFE_INTEGER;
-	var yMax = 0;
+    //find min and max
+    var yMin = Number.MAX_SAFE_INTEGER;
+    var yMax = 0;
 
-	var xMin = Number.MAX_SAFE_INTEGER;
-	var xMax = 0;
-	// if(sort == true){
-	// 	data.sort(function(a,b){
-	// 		return a - b;
-	// 	});
-	// }
+    var xMin = Number.MAX_SAFE_INTEGER;
+    var xMax = 0;
+    // if(sort == true){
+    // 	data.sort(function(a,b){
+    // 		return a - b;
+    // 	});
+    // }
 
-	yData.forEach(function(d){
-		yMin = Math.min(yMin, d);
-		yMax = Math.max(yMax, d);
-	});
+    yData.forEach(function(d){
+	yMin = Math.min(yMin, d);
+	yMax = Math.max(yMax, d);
+    });
 
-	xData.forEach(function(d){
-		xMin = Math.min(xMin, d);
-		xMax = Math.max(xMax, d);
-	});
+    xData.forEach(function(d){
+	xMin = Math.min(xMin, d);
+	xMax = Math.max(xMax, d);
+    });
 
-	var leastSquaresCoeff = leastSquares(xData.slice(), yData.slice());
-	var regressionY = leastSquaresCoeff["y_res"];
-	var corre_coef = leastSquaresCoeff["corre_coef"];
-	//set scale
-	var xScale;
-	var yScale;
-	// var yScale = d3.scale.log().domain([minVal + 0.000001, maxVal + 0.000001]).range([height, 0]);:q
+    var leastSquaresCoeff = leastSquares(xData.slice(), yData.slice());
+    var regressionY = leastSquaresCoeff["y_res"];
+    var corre_coef = leastSquaresCoeff["corre_coef"];
+    //set scale
+    var xScale;
+    var yScale;
+    // var yScale = d3.scale.log().domain([minVal + 0.000001, maxVal + 0.000001]).range([height, 0]);:q
 
-	var svg;
+    var svg;
 
-	function visualize(){
+    function visualize(){
 
-		xScale = d3.scale.linear().domain([xMin, xMax]).range([0, width]);
-		yScale = d3.scale.linear().domain([yMin, yMax]).range([height, 0]);
+	xScale = d3.scale.linear().domain([xMin, xMax]).range([0, width]);
+	yScale = d3.scale.linear().domain([yMin, yMax]).range([height, 0]);
 
-		svg = d3.select(containerID).append('svg')
-				.attr('width', width + margin.left + margin.right)
-				.attr('height', height + margin.top + margin.bottom)
-				.append('g')
-					.attr('transform', "translate(" + margin.left + "," + margin.top + ")");
+	svg = d3.select(containerID).append('svg')
+	    .attr('width', width + margin.left + margin.right)
+	    .attr('height', height + margin.top + margin.bottom)
+	    .append('g')
+	    .attr('transform', "translate(" + margin.left + "," + margin.top + ")");
 
-		var xAxis = d3.svg.axis()
-					.scale(xScale)
-					.orient('bottom')
-					.ticks(5)
-					// .tickFormat(d3.formatPrefix(".1", 1e6));
-					.tickFormat(d3.format(".1e"));	
-		var yAxis = d3.svg.axis()
-					.scale(yScale)
-					.orient('left')
-					.ticks(5)
-					.tickFormat(d3.format(".1e"));		
+	var xAxis = d3.svg.axis()
+	    .scale(xScale)
+	    .orient('bottom')
+	    .ticks(5)
+	// .tickFormat(d3.formatPrefix(".1", 1e6));
+	    .tickFormat(d3.format(".1e"));	
+	var yAxis = d3.svg.axis()
+	    .scale(yScale)
+	    .orient('left')
+	    .ticks(5)
+	    .tickFormat(d3.format(".1e"));		
 
-		var xAxisLine = svg.append('g')
-						.attr('class', 'xAxis')
-						.attr("transform", "translate(0," + height + ")")
-						.call(xAxis)
-		xAxisLine.append('text')
-				.attr('x', width)
-				.attr('y', -6)
-				.style('text-anchor', 'end')
-				.text("Exclusive Runtime")
+	var xAxisLine = svg.append('g')
+	    .attr('class', 'xAxis')
+	    .attr("transform", "translate(0," + height + ")")
+	    .call(xAxis)
+	xAxisLine.append('text')
+	    .attr('x', width)
+	    .attr('y', -6)
+	    .style('text-anchor', 'end')
+	    .text("Exclusive Runtime")
 
-		var yAxisLine = svg.append('g')
-							.attr('class', 'yAxis')
-							.call(yAxis)
-		yAxisLine.append("text")
-				.attr('transform', 'rotate(-90)')
-				.attr('y', 6)
-				.attr('dy', ".71em")
-				.style("text-anchor", "end")
-				.text("Inclusive Runtime");
+	var yAxisLine = svg.append('g')
+	    .attr('class', 'yAxis')
+	    .call(yAxis)
+	yAxisLine.append("text")
+	    .attr('transform', 'rotate(-90)')
+	    .attr('y', 6)
+	    .attr('dy', ".71em")
+	    .style("text-anchor", "end")
+	    .text("Inclusive Runtime");
 
-		svg.append('g').selectAll('.dot')
-			.data(yData)
-			.enter().append('circle')
-			.attr('class', 'dot')
-			.attr('r', 3)
-			.attr('cx', function(d, i){
-				return xScale(xData[i]);
-			})
-			.attr('cy', function(d, i){
-				return yScale(d);
-			})
-			.style('fill', "#4682b4")
+	svg.append('g').selectAll('.dot')
+	    .data(yData)
+	    .enter().append('circle')
+	    .attr('class', 'dot')
+	    .attr('r', 3)
+	    .attr('cx', function(d, i){
+		return xScale(xData[i]);
+	    })
+	    .attr('cy', function(d, i){
+		return yScale(d);
+	    })
+	    .style('fill', "#4682b4")
 
-		var line = d3.svg.line()
-			        .x(function(d,i) {
-			            return xScale(xData[i]);
-			        })
-			        .y(function(d) {
-			            return yScale(d);
-			        });
+	var line = d3.svg.line()
+	    .x(function(d,i) {
+		return xScale(xData[i]);
+	    })
+	    .y(function(d) {
+		return yScale(d);
+	    });
 
-		var trendline = svg.append('g').append("path")
-	        .datum(regressionY)
-	        .attr("class", "res_line")
-	        .attr("d", line)
-	        .style("stroke", "black")
-			.style("stroke-width", "1px")
-			.style("opacity", 0.5);	
+	var trendline = svg.append('g').append("path")
+	    .datum(regressionY)
+	    .attr("class", "res_line")
+	    .attr("d", line)
+	    .style("stroke", "black")
+	    .style("stroke-width", "1px")
+	    .style("opacity", 0.5);	
 
-		var coefText = svg.append('g');
-		var decimalFormat = d3.format("0.2f");
-		coefText.append('text')
-				.text("corr-coef: " + decimalFormat(corre_coef))
-				.attr("x", function(d) {return width - 100;})
-				.attr("y", function(d) {return 10;});	
+	var coefText = svg.append('g');
+	var decimalFormat = d3.format("0.2f");
+	coefText.append('text')
+	    .text("corr-coef: " + decimalFormat(corre_coef))
+	    .attr("x", function(d) {return width - 100;})
+	    .attr("y", function(d) {return 10;});	
 
         xAxisLine.selectAll('path')
-				.style("fill", "none")
-				.style("stroke", "black")
-				.style("stroke-width", "1px");
-		xAxisLine.selectAll('line')
-				.style("fill", "none")
-				.style("stroke", "#000")
-				.style("stroke-width", "1px")
-				.style("opacity", 0.5);
+	    .style("fill", "none")
+	    .style("stroke", "black")
+	    .style("stroke-width", "1px");
+	xAxisLine.selectAll('line')
+	    .style("fill", "none")
+	    .style("stroke", "#000")
+	    .style("stroke-width", "1px")
+	    .style("opacity", 0.5);
         xAxisLine.selectAll("text")
-				.style('font-size', '10px')
-				.style('font-family', 'sans-serif')
-				.style('font-weight', 'lighter');		
+	    .style('font-size', '10px')
+	    .style('font-family', 'sans-serif')
+	    .style('font-weight', 'lighter');		
 
         yAxisLine.selectAll('path')
-				.style("fill", "none")
-				.style("stroke", "black")
-				.style("stroke-width", "1px");
-		yAxisLine.selectAll('line')
-				.style("fill", "none")
-				.style("stroke", "#000")
-				.style("stroke-width", "1px")
-				.style("opacity", 0.5);
+	    .style("fill", "none")
+	    .style("stroke", "black")
+	    .style("stroke-width", "1px");
+	yAxisLine.selectAll('line')
+	    .style("fill", "none")
+	    .style("stroke", "#000")
+	    .style("stroke-width", "1px")
+	    .style("opacity", 0.5);
         yAxisLine.selectAll("text")
-				.style('font-size', '10px')
-				.style('font-family', 'sans-serif')
-				.style('font-weight', 'lighter');						
-	}
+	    .style('font-size', '10px')
+	    .style('font-family', 'sans-serif')
+	    .style('font-weight', 'lighter');						
+    }
+
+    visualize();
+
+    this.setContainerWidth = function(newWidth){
+	containerWidth = newWidth;
+	width = containerWidth - margin.left - margin.right;
+
+	console.log(containerWidth);
+    }
+
+    this.setContainerHeight = function(newHeight){
+	containerHeight = newHeight;
+	height = containerHeight - margin.top - margin.bottom;
+    }
+
+    this.reDraw = function(){
+	$(containerID).empty();
 
 	visualize();
+    }
 
-	this.setContainerWidth = function(newWidth){
-		containerWidth = newWidth;
-		width = containerWidth - margin.left - margin.right;
-
-		console.log(containerWidth);
-	}
-
-	this.setContainerHeight = function(newHeight){
-		containerHeight = newHeight;
-		height = containerHeight - margin.top - margin.bottom;
-	}
-
-	this.reDraw = function(){
-		$(containerID).empty();
-
-		visualize();
-	}
-
-	// returns slope, intercept and r-square of the line
-	function leastSquares(xSeries, ySeries) {
-		var n = xSeries.length;
-		var x_mean = 0;
+    // returns slope, intercept and r-square of the line
+    function leastSquares(xSeries, ySeries) {
+	var n = xSeries.length;
+	var x_mean = 0;
         var y_mean = 0;
         var term1 = 0;
         var term2 = 0;
@@ -227,41 +227,41 @@ function Scatter(args){
 
         //compute correlation coef
         var xy = [];
-	    var x2 = [];
-	    var y2 = [];
-	  
-	    for(var i=0; i<n; i++) {
-	        xy.push(xSeries[i] * ySeries[i]);
-	        x2.push(xSeries[i] * xSeries[i]);
-	        y2.push(ySeries[i] * ySeries[i]);
-	    }
+	var x2 = [];
+	var y2 = [];
+	
+	for(var i=0; i<n; i++) {
+	    xy.push(xSeries[i] * ySeries[i]);
+	    x2.push(xSeries[i] * xSeries[i]);
+	    y2.push(ySeries[i] * ySeries[i]);
+	}
 
-		var sum_x = 0;
-	    var sum_y = 0;
-	    var sum_xy = 0;
-	    var sum_x2 = 0;
-	    var sum_y2 = 0;
-	  
-	    for(var i=0; i< n; i++) {
-	        sum_x += xSeries[i];
-	        sum_y += ySeries[i];
-	        sum_xy += xy[i];
-	        sum_x2 += x2[i];
-	        sum_y2 += y2[i];
-	    }
-		var step1 = (n * sum_xy) - (sum_x * sum_y);
-	    var step2 = (n * sum_x2) - (sum_x * sum_x);
-	    var step3 = (n * sum_y2) - (sum_y * sum_y);
-	    var step4 = Math.sqrt(step2 * step3);
-	    var corre_coef = step1 / step4;
+	var sum_x = 0;
+	var sum_y = 0;
+	var sum_xy = 0;
+	var sum_x2 = 0;
+	var sum_y2 = 0;
+	
+	for(var i=0; i< n; i++) {
+	    sum_x += xSeries[i];
+	    sum_y += ySeries[i];
+	    sum_xy += xy[i];
+	    sum_x2 += x2[i];
+	    sum_y2 += y2[i];
+	}
+	var step1 = (n * sum_xy) - (sum_x * sum_y);
+	var step2 = (n * sum_x2) - (sum_x * sum_x);
+	var step3 = (n * sum_y2) - (sum_y * sum_y);
+	var step4 = Math.sqrt(step2 * step3);
+	var corre_coef = step1 / step4;
 
-	    console.log(corre_coef);
+	console.log(corre_coef);
 
         return {
-        	"y_res" : yhat,
-        	"corre_coef" : corre_coef
+            "y_res" : yhat,
+            "corre_coef" : corre_coef
         };
 
-	}
+    }
 
 }
